+++ 4-full-mvc-framework/__install/sql_files/notes.sql
@@ -1,32 +1,14 @@
--- phpMyAdmin SQL Dump
--- version 3.5.6
--- http://www.phpmyadmin.net
---
--- Host: localhost
--- Erstellungszeit: 28. Jun 2013 um 01:13
--- Server Version: 5.5.29-log
--- PHP-Version: 5.3.21
-
 SET SQL_MODE="NO_AUTO_VALUE_ON_ZERO";
 SET time_zone = "+00:00";
-
 
 /*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
 /*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
 /*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
 /*!40101 SET NAMES utf8 */;
 
---
--- Datenbank: `login`
---
+-- Database: `login`
 
--- --------------------------------------------------------
-
---
--- Tabellenstruktur für Tabelle `note`
---
-
-CREATE TABLE IF NOT EXISTS `note` (
+CREATE TABLE IF NOT EXISTS `notes` (
   `note_id` int(11) NOT NULL AUTO_INCREMENT,
   `note_text` text NOT NULL,
   `user_id` int(11) NOT NULL,